+++ src/models/commandToggle.js
@@ -1,0 +1,25 @@
+// export function that defines commandToggle model
+module.exports = (sequelize, Sequelize) => {
+    return sequelize.define('commandToggle', {
+        id: {
+            type: Sequelize.INTEGER,
+            primaryKey: true,
+            autoIncrement: true
+        },
+        serverId: {
+            type: Sequelize.STRING(20),
+            allowNull: false,
+            unique: 'unique_toggle'
+        },
+        commandName: {
+            type: Sequelize.STRING(32),
+            allowNull: false,
+            unique: 'unique_toggle'
+        },
+        enabled: {
+            type: Sequelize.BOOLEAN,
+            allowNull: false,
+            defaultValue: true
+        }
+    });
+};