+++ src/syncdb.js
@@ -2,7 +2,7 @@ require('dotenv').config();
 const sequelize = require('./utilities/database');
 
 
-const { Guild, Giveaway, Question, Application, StaffRoles, CustomRoles, Embed, Level, LevelRoles, XPSettings, XPIgnoredChannels } = require('./models/')
+const { Guild, Giveaway, Question, Application, StaffRoles, CustomRoles, Embed, Level, LevelRoles, XPSettings, XPIgnoredChannels, Birthday, CommandToggle } = require('./models/')
 
 Guild.sync();
 Giveaway.sync();
@@ -15,3 +15,7 @@ Level.sync();
 LevelRoles.sync();
 XPSettings.sync();
 XPIgnoredChannels.sync();
+Birthday.sync();
+CommandToggle.sync();
+
+console.log('All models were synchronized successfully.');